cleaning up code and files
- making more files and renaming things
