--- algorithm/sandbox/2DMatrix/desc.json
@@ -1,8 +1,0 @@
-{
-  "2DMatrix": "2DMatrix for drag and drop tracers!",
-  "References": [
-  ],
-  "files": {
-    "normal": "2D Matrix"
-  }
-}
