+++ algorithm/sandbox/create_mode/desc.json
@@ -1,0 +1,8 @@
+{
+  "create_mode": "Words have to go here",
+  "References": [
+  ],
+  "files": {
+    "normal": "Create Mode"
+  }
+}
